More fixes for NPP 8.3.3 plugin API compatibility
Change more places to (u)intptr_t, and remove a rather obvious
static_cast<intptr_t> that prevented the previous commit to work
with files > 2GB.

+++ src/LuaConsole.cpp
@@ -363,7 +363,7 @@ void LuaConsole::showAutoCompletion() {
 
 		// Back up past the partial word
 		prevCh = static_cast<int>(m_sciInput->Call(SCI_GETCHARAT, curPos - 1 - partialWord.size()));
-		curPos = curPos - static_cast<int>(partialWord.size());
+		curPos = curPos - partialWord.size();
 	}
 
 	if (prevCh == '.' || prevCh == ':') {
